applications/MQTT/requirements.txt
paho-mqtt
valkey
PyYAML